Organize server scripts and remove unused code
