+++ server/scripts/migrate.sql
