+++ server/scripts/update-db.ts
@@ -1,4 +1,4 @@
-import { db } from './db';
+import { db } from '../db';
 import { sql } from 'drizzle-orm';
 
 async function updateDatabase() {
